third_party/WebKit/LayoutTests/fast/notifications/notifications-permission.html

<!DOCTYPE html>
<html>
<head>
	<script src="../../resources/js-test.js"></script>
</head>

<body>
<div id="description"></div>
<div id="console"></div>

<script type="text/javascript">
description("Check for existence of Notification.permission");

shouldBeDefined("window.Notification");
shouldBeDefined("window.Notification.permission");

var successfullyParsed = true;
</script>
</body>
</html>
